moving some integration tests in auto-configure package
