+++ ci-droid-tasks-consumer-autoconfigure/src/test/resources/application-bitbutcket-synchronous-test.yml
